fix long page loading
The reason behind longer loading were wrong usage of variables.
const loaded = sub1.ready() && sub2.ready()
this is a reactive variable and it will first create instance of sub1
and wait for it .ready() callback and the create instance of sub2 and
wait for it's ready() callback
So every subscription must be created in separate variable.